Update conditions for dev/main branches and preview build config
for PRs against `main` we still want to have the preview functionality.
however a "development" build is not the same as a "production" build so
we don't want to take the risk and do that!

To elaborate, "the risk" is that we build a development preview build,
maybe test it directly from the PR preview URL, instead of building it
locally and then declaring it PASS, while in reality a production build
may be failing!

IMO the above mentioned risk is extremely small in practice, however in theory it isn't zero!

This change renders the PR preview build impossible to connect to the
CreditcoinLocal network which eliminates the risk of testing one thing
but shipping another! The inconvenience is that any testing on PRs
against `main` branch will require also testing against Devnet (or
Testnet, or Mainnet)!